% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{dragon_curve}
\alias{dragon_curve}
\title{Dragon Curve}
\usage{
dragon_curve(order)
}
\arguments{
\item{order}{The order of the dragon curve, i.e. how many times it will be
folded.}
}
\value{
A matrix with two columns for the x and y coordinates of the folds.
}
\description{
Repeatedly fold a line to create a space-filling curve.
}
\examples{
d <- dragon_curve(4)
d
plot(d)

}
\references{
\url{https://en.wikipedia.org/wiki/Dragon_curve}
}
