% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{fold_dragon}
\alias{fold_dragon}
\title{Fold the Dragon Curve}
\usage{
fold_dragon(order, limit = 20)
}
\arguments{
\item{order}{The order of the dragon curve. Must be a positive
integer <= \code{limit}.}

\item{limit}{The maximum allowed value for \code{order}. This is to prevent
accidental creation of very big results which can consume cpu and memory.
Defaults to 20. Be aware that the resulting vector will have a length of
2^\code{order}-1 and the coordinate matrix returned by
\code{get_path_coordinates} will be twice that object size.}
}
\value{
A numeric vector of length 2^\code{order}-1. The values are either
  1 for left turns or -1 for right turns.
}
\description{
\code{fold_dragon} creates a vector which indicates in which direction each
corner of the dragon curve turns.
}
\details{
This function is supposed to be used in combination with
\code{\link{get_path_coordinates}} for which it creates the input.
}
\examples{
fold_dragon(1)
fold_dragon(3)
fold_dragon(5)
}
\seealso{
Other dragon curve functions: \code{\link{flip_dragon}},
  \code{\link{get_path_coordinates}},
  \code{\link{rotate_dragon}}
}
\concept{dragon curve functions}
