% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractalplotr-package.R
\docType{package}
\name{fractalplotr-package}
\alias{fractalplotr}
\alias{fractalplotr-package}
\title{fractalplotr: Plot Beautiful Fractals with R}
\description{
Plot beautiful fractals in R. This package lets you
    create fractals of various types, for example mandelbrot sets, dragon
    curves or sandpiles.
}
\details{
TODO: add details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BastiHz/fractalplotr}
  \item Report bugs at \url{https://github.com/BastiHz/fractalplotr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Henz \email{bastihz.dev@posteo.de}

}
