% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractalplotr.R
\docType{package}
\name{fractalplotr}
\alias{fractalplotr}
\alias{fractalplotr-package}
\title{fractalplotr: Plot Beautiful Fractals with R}
\description{
TODO: Add package documentation.
}
\section{This is a section}{

Test 123.
}

\section{This is another section}{

Test test 456.
}

