% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{get_path_coordinates}
\alias{get_path_coordinates}
\title{Calculate the Coordinates for the Corners of the Dragon Curve}
\usage{
get_path_coordinates(turns)
}
\arguments{
\item{turns}{A numeric vector consisting solely of the values 1 and -1. A 1
results in a left turn and a -1 makes a right turn.}
}
\value{
A numeric matrix with two columns x and y containing the coordinates
  of the folds in the dragon curve.
}
\description{
\code{get_path_coordinates} accepts a vector of turn directions and returns
the coordinates of the folds in the associated curve.
}
\details{
This function is supposed to be used in combination with
\code{\link{fold_dragon}}. Of course, it can also return a path for other
vectors of turns if they conform to the input format.
}
\examples{
folds <- fold_dragon(3)
get_path_coordinates(folds)

# The matrix can be used easily plot the dragon curve:
folds <- fold_dragon(8)
coords <- get_path_coordinates(folds)
plot(coords, type = "l", asp = 1)
}
\seealso{
Other dragon curve functions: \code{\link{flip_dragon}},
  \code{\link{fold_dragon}}, \code{\link{rotate_dragon}}
}
\concept{dragon curve functions}
