% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror}
\alias{mirror}
\alias{mirror.default}
\alias{mirror.dragon_curve}
\title{Mirror Fractals}
\usage{
mirror(fractal, direction)

\method{mirror}{default}(fractal, direction = c("horizontal",
  "vertical"))

\method{mirror}{dragon_curve}(fractal, direction = c("horizontal",
  "vertical"))
}
\arguments{
\item{fractal}{A fractal object to be mirrored}

\item{direction}{The direction of the mirroring, either "horizontal" or
"vertical". Can be abbreviated.}
}
\value{
The mirrored fractal.
}
\description{
Mirror fractals horizontally or vertically.
}
\examples{
d <- dragon_curve(10)
d_h <- mirror(d, "horizontal")
d_v <- mirror(d, "vertical")
plot(d)
plot(d_h)
plot(d_v)

}
