% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_color_matrix.R
\name{plot.color_matrix}
\alias{plot.color_matrix}
\title{Plot color matrices}
\usage{
\method{plot}{color_matrix}(x, ...)
}
\arguments{
\item{x}{A matrix of color values.}

\item{...}{Not used.}
}
\value{
None
}
\description{
Plot objects of class "color_matrix" as returned from e.g. \code{\link[=mandelbrot]{mandelbrot()}} or
\code{\link[=sandpile]{sandpile()}}.
}
\examples{
plot(mandelbrot(width = 200, height = 150))

}
