% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fractal}
\alias{plot_fractal}
\alias{plot_fractal.color_matrix}
\alias{plot_fractal.dragon_curve}
\title{Plot fractals}
\usage{
plot_fractal(fractal, ...)

\method{plot_fractal}{color_matrix}(color_matrix)

\method{plot_fractal}{dragon_curve}(dragon, color = "black")
}
\arguments{
\item{fractal}{The fractal to be plotted.}

\item{color}{The color of the dragon curve as single value in a format
that R understands.}
}
\description{
Plot fractals
}
\examples{
plot_fractal(dragon_curve(12), "orange")
}
