% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radians.R
\name{radians}
\alias{radians}
\title{Radians}
\usage{
radians(degrees)
}
\arguments{
\item{degrees}{The degrees to convert to radians.}
}
\value{
The radians equivalent to the input degrees.
}
\description{
Convert degrees to radians.
}
\details{
This function just returns \code{degrees * pi / 180}.
}
\examples{
radians(90)

}
