% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\alias{rotate.matrix}
\alias{rotate.dragon_curve}
\title{Rotate Fractals}
\usage{
rotate(fractal, n)

\method{rotate}{matrix}(fractal, n)

\method{rotate}{dragon_curve}(fractal, n)
}
\arguments{
\item{fractal}{The fractal to be rotated.}

\item{n}{An integer specifying the number and direction of turns. A positive
value turns counterclockwise and a negative value turns clockwise.}
}
\value{
The rotated fractal.
}
\description{
Rotate a fractal by multiples of 90 degrees.
}
\examples{
d <- dragon_curve(10)
d_l <- rotate(d, 1)  # rotate once to the left
d_r <- rotate(d, -2)  # rotate twice to the right
plot(d)
plot(d_l)
plot(d_r)

}
