% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{rotate_dragon}
\alias{rotate_dragon}
\title{Rotate a Dragon Curve}
\usage{
rotate_dragon(coordinates, direction, times = 1)
}
\arguments{
\item{coordinates}{A matrix with two columns for the x and y coordinates of
the folds in the curve. Usually created by
\code{\link{get_path_coordinates}}.}

\item{direction}{A string specifying the direction of rotation. Possible
values are "left", "right", "clockwise" and "anticlockwise".}

\item{times}{A single positive number spcifying how often to rotate by 90°.}
}
\value{
A matrix containing the rotated coordinates.
}
\description{
Rotate dragon curve coordinates 90° to the right (clockwise) or left
(anticlockwise) once or multiple times.
}
\examples{
ragon_curve <- get_path_coordinates(fold_dragon(2))
rotate_dragon(dragon_curve, "left", 1)
}
\seealso{
Other dragon curve functions: \code{\link{flip_dragon}},
  \code{\link{fold_dragon}},
  \code{\link{get_path_coordinates}}
}
\concept{dragon curve functions}
