% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandpile.R
\name{sandpile}
\alias{sandpile}
\title{Title}
\usage{
sandpile(height, colors = c(1, 2/3, 1/3, 0), return_colors = TRUE)
}
\arguments{
\item{height}{foo}

\item{colors}{bar}

\item{return_colors}{baz}
}
\value{
foo bar color_matrix obj
}
\description{
TODO: fixme
}
\examples{
sandpile(100, return_colors = FALSE)
}
