% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_fractal}
\alias{save_fractal}
\alias{save_fractal.color_matrix}
\alias{save_fractal.dragon_curve}
\title{Save a fractal as an image}
\usage{
save_fractal(fractal, ...)

\method{save_fractal}{color_matrix}(color_matrix, filename)

\method{save_fractal}{dragon_curve}(dragon, color, filename, ...)
}
\arguments{
\item{fractal}{The fractal to be saved.}

\item{...}{Further arguments to \code{png()} or \code{plot_fractals()}}

\item{filename}{A character string naming a file path.}
}
\description{
TODO: Fill in description of the methods for color_matrix and dragon_curve.
}
