% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{dragon_curve}
\alias{dragon_curve}
\title{Dragon Curve}
\usage{
dragon_curve(order)
}
\arguments{
\item{order}{foo bar}
}
\value{
foo bar
}
\description{
TODO: this documentation
}
\details{
The length of the curve will be 2 ^ order - 1.
}
\examples{
dragon_curve(5)
}
