% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{dragon_curve}
\alias{dragon_curve}
\title{Dragon Curve}
\usage{
dragon_curve(n)
}
\arguments{
\item{n}{How often the curve will be folded. Must be at least 1.}
}
\value{
A matrix of class "dragon_curve" with two columns for the x and y
coordinates of the folds.
}
\description{
Repeatedly fold a line to create a space-filling curve.
}
\examples{
d <- dragon_curve(9)
plot(d)

}
\seealso{
\code{\link[=plot.dragon_curve]{plot.dragon_curve()}}
}
