% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{flip_dragon}
\alias{flip_dragon}
\title{Flip a Dragon Curve}
\usage{
flip_dragon(dragon, direction)
}
\arguments{
\item{dragon}{The dragon curve returned by \code{dragon_curve()}.}

\item{direction}{A string specifying the direction of the flip. Possible
values are "horizontal" and "vertical".}
}
\value{
The mirrored dragon curve coordinates.
}
\description{
Mirror a dragon curve either horizontally or vertically.
}
\examples{
d <- dragon_curve(3)
flip_dragon(d, "horizontal")
}
\seealso{
Other dragon curve functions: \code{\link{rotate_dragon}}
}
\concept{dragon curve functions}
