% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{flip_dragon}
\alias{flip_dragon}
\title{Flip a Dragon Curve}
\usage{
flip_dragon(coordinates, direction)
}
\arguments{
\item{coordinates}{A matrix with two columns for the x and y coordinates of
the folds in the curve. Usually created by
\code{\link{get_path_coordinates}}.}

\item{direction}{A string specifying the direction of the flip. Possible
values are "horizontal" and "vertical".}
}
\value{
The mirrored dragon curve coordinates.
}
\description{
Mirror a dragon curve either horizontally or vertically.
}
\examples{
dragon_curve <- get_path_coordinates(fold_dragon(3))
flip_dragon(dragon_curve, "horizontal")
}
\seealso{
Other dragon curve functions: \code{\link{fold_dragon}},
  \code{\link{get_path_coordinates}},
  \code{\link{rotate_dragon}}
}
\concept{dragon curve functions}
