% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandelbrot.R
\name{mandelbrot}
\alias{mandelbrot}
\title{Mandelbrot set}
\usage{
mandelbrot(width, height, re_width = 3.5, im_height = NA,
  center = complex(real = -0.5, imaginary = 0), max_iterations = 128,
  threshold = 2, return_colors = TRUE, color_palette = NULL,
  color_inside = "black", color_mode = c("simple", "histogram",
  "smooth"))
}
\arguments{
\item{width, height}{The width and height in pixels.}

\item{re_width, im_height}{The width and height of the complex plane}

\item{center}{A complex number giving the center of the complex plane.
Defaults to -0.5+0i.}

\item{max_iterations}{The maximum number of iterations. Defaults is 128}

\item{threshold}{The threshold value. Default is 2.}

\item{return_colors}{Logical. Should the colors or the number of steps be
returned?}

\item{color_palette}{A vector of colors. Should be the same length as
max_iterations.}

\item{color_inside}{The color of the area inside the set.}

\item{color_mode}{How the colors of the set will be calculated. One of
"simple", "histogram" or "smooth". Can be abbreviated.}
}
\value{
A matrix specifying the color for each coordinate.
}
\description{
The famous mandelbrot set.
}
\examples{
m <- mandelbrot(width = 200, height = 150)
plot(m)

}
\references{
\url{https://en.wikipedia.org/wiki/Mandelbrot_set}
}
