% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandelbrot.R
\name{mandelbrot}
\alias{mandelbrot}
\title{Title}
\usage{
mandelbrot(width, height, re_width = 3.5, im_height = NA,
  center = complex(real = -0.5, imaginary = 0), max_iterations = 128,
  threshold = 2, return_colors = TRUE, color_palette = NULL,
  color_inside = "black", color_mode = c("simple", "histogram",
  "smooth"))
}
\arguments{
\item{width}{foo}

\item{height}{bar}

\item{re_width}{baz}

\item{im_height}{hurr}

\item{center}{blubb}

\item{max_iterations}{hurz}

\item{threshold}{a}

\item{return_colors}{b}

\item{color_palette}{c}

\item{color_inside}{d}

\item{color_mode}{Can be abbreviated.}
}
\value{
waaa color_matrix
}
\description{
TODO: fixme
}
\examples{
mandelbrot(width = 10, height = 5)
}
