% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_system.R
\name{plot.l_system}
\alias{plot.l_system}
\title{Plot L-systems}
\usage{
\method{plot}{l_system}(x, ...)
}
\arguments{
\item{x}{A data frame of class "l_system" as returned from
\code{\link[=l_system]{l_system()}} with the columns x0, y0, x1, and y1.}

\item{...}{Other parameters passed on to \code{\link[graphics:segments]{graphics::segments()}}.}
}
\value{
None
}
\description{
Plot L-systems as line segments.
}
\examples{
L <- l_system(
    "X",
    list(`X` = "[@.7071-FX][@.7071+FX]"),
    n = 10,
    angle = pi * 0.2
)
plot(L)

}
