% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_system.R
\name{plot.l_system}
\alias{plot.l_system}
\title{Plot L-systems}
\usage{
\method{plot}{l_system}(x, ...)
}
\arguments{
\item{x}{A data frame of class "l_system" as returned from
\code{\link[=l_system]{l_system()}} with the columns x0, y0, x1, and y1.}

\item{...}{Other parameters passed on to \code{\link[graphics:segments]{graphics::segments()}}.}
}
\value{
None
}
\description{
Plot L-systems as line segments.
}
\examples{
l_tree <- l_system(
    axiom = "X",
    rules = list(`X` = "F[+@.7X]F![-@.6X]F"),
    n = 10,
    angle = radians(22.5),
    extra_info = TRUE
)
# using the extra_info argument to vary color and line thickness:
plot(
    l_tree,
    col = ifelse(l_tree$depth < 6, "sienna", "forestgreen"),
    lwd = l_tree$depth / max(l_tree$depth) * -2 + 3
)

}
