% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fractal}
\alias{plot_fractal}
\alias{plot.dragon_curve}
\alias{plot.color_matrix}
\title{Plot fractals}
\usage{
\method{plot}{dragon_curve}(x, pretty = TRUE, ...)

\method{plot}{color_matrix}(x, ...)
}
\arguments{
\item{x}{A fractal object.}

\item{pretty}{TODO: Explain what this does exactly. Ant to set it to false if
the user wants to customize the plot.}

\item{...}{Futher arguments passed on to \code{plot()}. Only used for dragon
curves.}
}
\description{
Plot fractals
}
\examples{
plot(dragon_curve(12), col = "orange")
}
