% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate Fractals}
\usage{
rotate(fractal, direction)
}
\arguments{
\item{fractal}{The fractal to be rotated.}

\item{direction}{The direction of rotation, either "clockwise" or
"counterclockwise". Can be abbreviated.}
}
\value{
The rotated fractal.
}
\description{
Rotate a fractal either clockwise or counterclockwise
}
\examples{
d1 <- dragon_curve(10)
d2 <- rotate(d, "clockwise")
plot(d1)
plot(d2)
}
