% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\alias{rotate.default}
\alias{rotate.dragon_curve}
\title{Rotate Fractals}
\usage{
rotate(fractal, n)

\method{rotate}{default}(fractal, n = 1)

\method{rotate}{dragon_curve}(fractal, n = 1)
}
\arguments{
\item{fractal}{The fractal to be rotated.}

\item{n}{A positive or negative integer specifying the number and direction
of 90 degree turns. A positive value turns clockwise and a negative value
turns counterclockwise.}
}
\value{
The rotated fractal.
}
\description{
Rotate a fractal either clockwise or counterclockwise by multiples of 90
degrees.
}
\examples{
d <- dragon_curve(10)
d_r <- rotate(d, 1)  # rotate once to the right
d_l <- rotate(d, -1)  # rotate once to the left
plot(d)
plot(d_r)
plot(d_l)

}
