% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_curve.R
\name{rotate_dragon}
\alias{rotate_dragon}
\title{Rotate a Dragon Curve}
\usage{
rotate_dragon(dragon, direction, times = 1)
}
\arguments{
\item{dragon}{A matrix with two columns for the x and y coordinates of
the folds in the curve. Usually created by
\code{\link{dragon_curve}}.}

\item{direction}{A string specifying the direction of rotation. Possible
values are "left", "right", "clockwise" and "counterclockwise".}

\item{times}{A single positive number specifying how often to rotate by 90°.}
}
\value{
A matrix containing the rotated coordinates.
}
\description{
Rotate dragon curve coordinates 90° to the right (clockwise) or left
(counterclockwise) once or multiple times.
}
\examples{
d <- dragon_curve(fold_dragon(3))
rotate_dragon(d, "left")
}
\seealso{
Other dragon curve functions: \code{\link{flip_dragon}}
}
\concept{dragon curve functions}
