% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandpile.R
\name{sandpile}
\alias{sandpile}
\title{Sandpile}
\usage{
sandpile(n, colors = c("white", "lightgray", "darkgray", "black"))
}
\arguments{
\item{n}{The number of grains dropped in the center.}

\item{colors}{A vector of length 4 or \code{NULL}. Specifies the colors to be
allocated to the values \code{0:4}. If \code{NULL}, the number of grains
per cell are returned instead.}
}
\value{
A matrix, either containing colors or the number of grains per cell,
  depending on the argument \code{colors}.
}
\description{
Drop some sand in the center and see it spread outwards.
}
\examples{
s <- sandpile(1000)
plot(s)

}
\references{
\url{https://en.wikipedia.org/wiki/Abelian_sandpile_model}
}
