% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandpile.R
\name{sandpile}
\alias{sandpile}
\title{Sandpiles}
\usage{
sandpile(n, colors = c("white", "lightgray", "darkgray", "black"),
  return_colors = TRUE)
}
\arguments{
\item{n}{The number of grains dropped in the center.}

\item{colors}{A vector of length 4 containing the colors to be allocated to
zero to three grains per cell.}

\item{return_colors}{Logical. Should the returned matrix consist of the
numbers of grains per cell or of the corresponding colors?}
}
\value{
A matrix. Either with the number of grains per cell or the
  corresponding colors, depending on the value of \code{return_colors}.
}
\description{
Drop some sand in the center and see it topple.
}
\examples{
sandpile(100, return_colors = FALSE)
}
\references{
\link{https://en.wikipedia.org/wiki/Abelian_sandpile_model}
}
